backend/requirements.txt
easydict==1.9
PyYAML==5.1.2
PyJWT==1.7.1
Flask==2.1.0
Flask-Cors==3.0.7
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Session==0.3.2
#Cython==0.29.13
redis==3.5.3
pandas==1.3.0
torch==1.8.0
torchvision==0.9.0
numpy
requests==2.22.0
pg8000==1.16.0
SQLAlchemy==1.3.18
uwsgi==2.0.18
itsdangerous==2.0.1
Werkzeug==2.0.0
Authlib==1.1.0